move filter_model_dict and dict_to_uuid to new bia_object_creation_utils.py file
