+++ bia-ingest/bia_ingest/ingest/biosample.py
@@ -1,12 +1,12 @@
 import logging
 from typing import List, Any, Dict, Optional
 
+from ..bia_object_creation_utils import dict_to_uuid, filter_model_dictionary
+
 from ..cli_logging import log_model_creation_count
 from .utils import (
     dicts_to_api_models,
     find_sections_recursive,
-    dict_to_uuid,
-    filter_model_dictionary,
 )
 from .biostudies import (
     Submission,
