+++ bia-ingest/bia_ingest/representation_creation/experimentally_captured_image.py
@@ -3,12 +3,13 @@ from uuid import UUID
 from typing import List, Dict, Any
 from pydantic import ValidationError
 
+from ..bia_object_creation_utils import filter_model_dictionary
+
 from ..cli_logging import log_failed_model_creation, log_model_creation_count
 
 from .utils import merge_dicts
-from ..ingest.utils import (
+from ..bia_object_creation_utils import (
     dict_to_uuid,
-    filter_model_dictionary,
 )
 from ..ingest.biostudies import (
     Submission,
